refactor(runtime): keep the delete-migration helper in the SlotList sibling
Rebasing onto main (which now carries #9001's sibling split) put shapes.rs
at 2013 lines, over the 2000-line cap. shape_index_migrate_after_delete is
mostly SlotList manipulation, so it moves next to retain_shift.
